cache tests
